42cursus/Common_Core/14-ft_transcendence/project/services/app/python/conf/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==24.4.2
Automat==22.10.0
certifi==2024.8.30
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
dj-rest-auth==6.0.0
Django==4.2.14
django-allauth==64.2.1
django-cors-headers==4.4.0
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
iniconfig==2.0.0
packaging==24.1
phonenumbers==8.13.45
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
pypng==0.20220715.0
pytest==8.3.2
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==1.26.15
zope.interface==6.4.post2
gunicorn==23.0.0
django-prometheus==2.3.1
